fix(scout-for-lol): repair Temporal verification contracts
