+++ packages/scout-for-lol/packages/backend/package.json
@@ -74,7 +74,6 @@
     "@temporalio/worker": "1.22.0",
     "@trpc/server": "11",
     "ai": "^7.0.77",
-    "cron": "^4.1.4",
     "cron-parser": "^5.5.0",
     "date-fns": "^4.1.0",
     "discord.js": "^14.26.4",
