+++ packages/scout-for-lol/packages/backend/src/betting/parlay-generation-types.ts
@@ -1,0 +1,13 @@
+import type {
+  LoadingScreenData,
+  PlayerConfigEntry,
+  QueueType,
+  RawCurrentGameInfo,
+} from "@scout-for-lol/data";
+
+export type StartParlayGenerationInput = {
+  gameInfo: RawCurrentGameInfo;
+  trackedPlayers: readonly PlayerConfigEntry[];
+  queueType: QueueType | undefined;
+  loadingScreenData: LoadingScreenData | undefined;
+};
