+++ packages/scout-for-lol/packages/docs-site/package.json
@@ -34,7 +34,6 @@
     "@astrojs/check": "^0.9.10",
     "@scout-for-lol/backend": "workspace:*",
     "@scout-for-lol/report": "workspace:*",
-    "@shepherdjerred/temporal": "workspace:*",
     "@shepherdjerred/eslint-config": "workspace:*",
     "@types/bun": "1.4.0",
     "@typescript/native": "npm:typescript@7.0.2",
